# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Raqib Api",
        "description": "API documentation for My App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/attendance/check-in": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Employee check-in",
                "operationId": "a6c74d0bf2e287e7e9e74a7237a72163",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checked in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "branch": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: IP or location not allowed"
                    },
                    "404": {
                        "description": "Not found: Employee or branch"
                    },
                    "409": {
                        "description": "Already clocked in"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/attendance/check-out": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Employee check-out",
                "operationId": "652ffef28ceb0c3bd7655251aa2be123",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checked out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "branch": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: IP or location not allowed or outside shift hours"
                    },
                    "404": {
                        "description": "Not found: Missing employee or branch"
                    },
                    "409": {
                        "description": "No open attendance record"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}